refactor: Refactor email templates
