+++ app/emails/verify.email.server.tsx
@@ -10,7 +10,7 @@ import {
 	Tailwind,
 	Text,
 } from '@react-email/components'
-import tailwindConfig from '../../../tailwind.config'
+import tailwindConfig from '../../tailwind.config'
 
 interface VerifyEmailProps {
 	otp: string
